requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
yfinance>=0.2.40
pandas>=2.0.0
pandas_ta>=0.3.14b
huggingface_hub>=0.23.0
llama-cpp-python>=0.2.80
requests>=2.31.0
